backend/foodgram/requirements.txt
asgiref==3.5.2
attrs==22.1.0
colorama==0.4.6
Django==4.1.3
django-filter==22.1
djangorestframework==3.14.0
exceptiongroup==1.0.4
flake8==5.0.4
iniconfig==1.1.1
isort==5.10.1
mccabe==0.7.0
packaging==21.3
Pillow==9.3.0
pluggy==1.0.0
pycodestyle==2.9.1
pyflakes==2.5.0
pyparsing==3.0.9
pytest==7.2.0
pytz==2022.6
sqlparse==0.4.3
tomli==2.0.1
tzdata==2022.6
